Add permissions script
