+++ challengeutils/__init__.py
@@ -3,4 +3,4 @@ from __future__ import absolute_import
 from . import createchallenge
 from . import mirrorwiki
 from . import utils
-
+from . import permissions
